ros666::drivers::block
Module sdcard

SDCard Driver

References:

Copied from rCore-Tutorial-v3/os/src/drivers/block/sdcard.rs

Author: wyfcyx

License: GPL-3.0 license

Modified by Gerrnperl

Dependencies:

Gerrnperl/k210-hal

(forked from wyfcyx/k210-hal)

(forked from riscv-rust/k210-hal)

wyfcyx/k210-pac

(forked from riscv-rust/k210-pac)

Gerrnperl/k210-soc

(forked from wyfcyx/k210-soc)

License of the dependencies:

Copyright 2019 RISC-V team

Permission to use, copy, modify, and/or distribute this software for any purpose with or without fee is hereby granted, provided that the above copyright notice and this permission notice appear in all copies.

THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.

Structs

PERIPHERALS 🔒

SDCard

SDCardCID

Card Identification Data: CID Register

SDCardCSD

Card Specific Data: CSD Register

SDCardInfo

Card information

SDCardWrapper

Enums

CMD

SD commands

InitError

Constants

SD_CS 🔒

CS value passed to SPI controller, this is a dummy value as SPI0_CS3 is not mapping to anything in the FPIOA

SD_CS_GPIONUM 🔒

GPIOHS GPIO number to use for controlling the SD card CS pin

SD_START_DATA_MULTIPLE_BLOCK_READ

Data token start byte, Start Multiple Block Read

SD_START_DATA_MULTIPLE_BLOCK_WRITE

Data token start byte, Start Multiple Block Write

SD_START_DATA_SINGLE_BLOCK_READ

Data token start byte, Start Single Block Read

SD_START_DATA_SINGLE_BLOCK_WRITE

Data token start byte, Start Single Block Write

SEC_LEN

Functions

init_sdcard 🔒

io_init 🔒

Connect pins to internal functions